% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLDC.R
\name{plotLDC}
\alias{plotLDC}
\title{Plot Load Duration Curve}
\usage{
plotLDC(x)
}
\arguments{
\item{x}{A data frame containing a field labeled "observed.loading", "loading.capacity", and "flow.percentile". Generally used on a site level.}
}
\description{
This function creates a plot of observed and loading capacities as a function of flow percentile for a given site.
}
