% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRecSeasonLoads.R
\name{plotRecSeasonLoads}
\alias{plotRecSeasonLoads}
\title{Plot Observed Loading and Loading Capacity over the Rec Season by Year}
\usage{
plotRecSeasonLoads(x, wndws = FALSE)
}
\arguments{
\item{x}{A data frame containing a field labeled "Observed_Loading", "Loading_Capacity", "Loading_Capacity_MOS", "Season", and "Flow_Percentile". Generally used on a site level.}

\item{wndws}{Logical. If TRUE, will create plots in individual windows. Default is false for Shiny.}
}
\description{
This function creates a barplot of observed and loading capacities by rec season for a given site.
}
