% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTimeSeries.R
\name{plotTimeSeries}
\alias{plotTimeSeries}
\title{Plot Parameter Concentrations Over Time}
\usage{
plotTimeSeries(x, max_crit, sitepal = "basel")
}
\arguments{
\item{x}{A data frame containing a field labeled "E.coli" and another labeled "Date", which has been converted to a date format.}

\item{max_crit}{Numeric. The maximum criterion for the E.coli dataset, taken from R317-2-14.}
}
\description{
This function creates a plot of parameter concentrations over time. Can handle plotting multiple sites over time.
}
