% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runTMDLApp.R
\name{runTMDLApp}
\alias{runTMDLApp}
\title{Run TMDL Shiny App Browser}
\usage{
runTMDLApp(wb_path, specs = TRUE, geom_crit = 126, max_crit = 406,
  mos = 0.1)
}
\arguments{
\item{wb_path}{A file path to the .xlsx file containing E.coli and flow data, linked by MLID/ML_Name/Date, contained in separate worksheets.}

\item{specs}{Logical. If TRUE, uses geom_crit, max_crit, and mos from workbook. If FALSE, function requires inputs of geom_crit, max_crit, and mos.}

\item{geom_crit}{Numeric. The geometric mean criterion for the E.coli dataset, taken from R317-2-14.}

\item{max_crit}{Numeric. The maximum criterion for the E.coli dataset, taken from R317-2-14.}

\item{mos}{Numeric proportion. The percent margin of safety (as a proportion) to use when calculating percent exceedance/reduction needed.}
}
\description{
Function uses output from calcLoadings to plot time series, LDC, and monthly and seasonal geomeans in shiny browser.
}
