% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmdlCalcs.R
\name{tmdlCalcs}
\alias{tmdlCalcs}
\title{Prep data for TMDL analysis and plotting}
\usage{
tmdlCalcs(wb_path, exportfromfunc = FALSE)
}
\arguments{
\item{wb_path}{A file path to the .xlsx file containing E.coli and flow data, linked by MLID/ML_Name/Date, contained in separate worksheets.}

\item{exportfromfunc}{Logical. Indicates whether workbook should be exported from tmdlCalcs function, or skipped if using Shiny interface. Default is FALSE to accommodate Shiny use.}
}
\value{
The output includes a new Excel workbook with the name of the original file plus today's date.xlsx, as well as the following dataframes, composed within a list: ecoli concentrations, flow data, ldc data, monthly geomeans, rec/non rec geomeans, and irg/non irg geomeans.
}
\description{
This function takes E.coli concentration and flow data (if applicable) to calculate geomean concentrations and loadings on a daily, monthly, rec-season, and irrigation season basis. Produces outputs that may be fed into plotting functions within the tmdlTools package.
}
