% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmdlCalcs.R
\name{tmdlCalcs}
\alias{tmdlCalcs}
\title{Prep data for TMDL analysis and plotting}
\usage{
tmdlCalcs(wb_path, specs = TRUE, rec_ssn = c("05-01", "10-31"),
  irg_ssn = c("05-15", "10-15"), geom_crit, max_crit, mos = 0.1,
  cf = 1000/100 * 28.3168 * 3600 * 24, overwrite = FALSE)
}
\arguments{
\item{wb_path}{A file path to the .xlsx file containing E.coli and flow data, linked by MLID/ML_Name/Date, contained in separate worksheets.}

\item{specs}{Logical. If TRUE, uses geom_crit, max_crit, and mos from workbook. If FALSE, function requires inputs of geom_crit, max_crit, and mos.}

\item{rec_ssn}{A character string defining the recreation season over which to calculate geometric means. Defaults to May 1 to October 31.}

\item{irg_ssn}{A character string defining the irrigation season over which to calculate geometric means. Defaults to May 15 to October 15.}

\item{geom_crit}{Numeric. The geometric mean criterion for the E.coli dataset, taken from R317-2-14.}

\item{max_crit}{Numeric. The maximum criterion for the E.coli dataset, taken from R317-2-14.}

\item{mos}{Numeric proportion. The percent margin of safety (as a proportion) to use when calculating percent exceedance/reduction needed.}

\item{cf}{Numeric. A correction factor to convert E.coli concentrations and flow data to loadings (amount per day). Default converts MPN/100 mL and cfs to MPN/day.}

\item{overwrite}{Logical. If TRUE, function updates input .xlsx file. If FALSE, function writes a new workbook with the name of the original file plus today's date.xlsx.}
}
\description{
This function takes E.coli concentration and flow data (if applicable) to calculate geomean concentrations and loadings on a daily, monthly, rec-season, and irrigation season basis. Produces outputs that may be fed into plotting functions within the tmdlTools package.
}
